+++ packages/payloadset/packages/xns/plugins/record/src/index.ts
@@ -1,2 +1,2 @@
 export * from './estimate/index.ts'
-export * from './helpers/index.ts'
+export * from './validation/index.ts'
