+++ packages/payloadset/packages/xns/plugins/record/src/validation/name/spec/tsconfig.json
