+++ packages/payloadset/packages/xns/plugins/record/src/validation/validation/validators/spec/tsconfig.json
